changes to circleci testing to also do in mem testing

+++ .circleci/doTests.sh
@@ -7,24 +7,23 @@ function cleanup {
 trap cleanup EXIT
 cleanup
 
-pinnedDBJson=`curl -s -X GET \
+pinnedDBJson=$(curl -s -X GET \
   'https://api.supertokens.io/0/plugin/pinned?planType=FREE' \
-  -H 'api-version: 0'`
-pinnedDBLength=`echo $pinnedDBJson | jq ".plugins | length"`
-pinnedDBArray=`echo $pinnedDBJson | jq ".plugins"`
+  -H 'api-version: 0')
+pinnedDBLength=$(echo "$pinnedDBJson" | jq ".plugins | length")
+pinnedDBArray=$(echo "$pinnedDBJson" | jq ".plugins")
 echo "got pinned dbs..."
 
-pluginInterfaceJson=`cat ../pluginInterfaceSupported.json`
-pluginInterfaceLength=`echo $pluginInterfaceJson | jq ".versions | length"`
-pluginInterfaceArray=`echo $pluginInterfaceJson | jq ".versions"`
+pluginInterfaceJson=$(cat ../pluginInterfaceSupported.json)
+pluginInterfaceLength=$(echo "$pluginInterfaceJson" | jq ".versions | length")
+pluginInterfaceArray=$(echo "$pluginInterfaceJson" | jq ".versions")
 echo "got plugin interface relations"
 
-coreDriverJson=`cat ../coreDriverInterfaceSupported.json`
-coreDriverLength=`echo $coreDriverJson | jq ".versions | length"`
-coreDriverArray=`echo $coreDriverJson | jq ".versions"`
+coreDriverJson=$(cat ../coreDriverInterfaceSupported.json)
+coreDriverArray=$(echo "$coreDriverJson" | jq ".versions")
 echo "got core driver relations"
 
-./getPluginInterfaceExactVersions.sh $pluginInterfaceLength "$pluginInterfaceArray"
+./getPluginInterfaceExactVersions.sh "$pluginInterfaceLength" "$pluginInterfaceArray"
 
 if [[ $? -ne 0 ]]
 then
@@ -35,7 +34,7 @@ else
 fi
 
 # get core version
-coreVersion=`cat ../build.gradle | grep -e "version =" -e "version="`
+coreVersion=$(cat ../build.gradle | grep -e "version =" -e "version=")
 while IFS='"' read -ra ADDR; do
     counter=0
     for i in "${ADDR[@]}"; do
@@ -47,7 +46,7 @@ while IFS='"' read -ra ADDR; do
     done
 done <<< "$coreVersion"
 
-responseStatus=`curl -s -o /dev/null -w "%{http_code}" -X PUT \
+responseStatus=$(curl -s -o /dev/null -w "%{http_code}" -X PUT \
   https://api.supertokens.io/0/core \
   -H 'Content-Type: application/json' \
   -H 'api-version: 0' \
@@ -57,8 +56,8 @@ responseStatus=`curl -s -o /dev/null -w "%{http_code}" -X PUT \
 	\"version\":\"$coreVersion\",
 	\"pluginInterfaces\": $pluginInterfaceArray,
 	\"coreDriverInterfaces\": $coreDriverArray
-}"`
-if [ $responseStatus -ne "200" ]
+}")
+if [ "$responseStatus" -ne "200" ]
 then
     echo "failed core PUT API status code: $responseStatus. Exiting!"
 	exit 1
@@ -71,56 +70,74 @@ do
         continue
     fi
     i=0
-    currTag=`echo $line | jq .tag`
-    currTag=`echo $currTag | tr -d '"'`
+    currTag=$(echo "$line" | jq .tag)
+    currTag=$(echo "$currTag" | tr -d '"')
 
-    currVersion=`echo $line | jq .version`
-    currVersion=`echo $currVersion | tr -d '"'`
+    currVersion=$(echo "$line" | jq .version)
+    currVersion=$(echo "$currVersion" | tr -d '"')
     piX=$(cut -d'.' -f1 <<<"$currVersion")
     piY=$(cut -d'.' -f2 <<<"$currVersion")
     piVersion="$piX.$piY"
     
-    while [ $i -lt $pinnedDBLength ]; do 
+    while [ $i -lt "$pinnedDBLength" ]; do
         someTestsRan=true
-        currPinnedDb=`echo $pinnedDBArray | jq ".[$i]"`
-        currPinnedDb=`echo $currPinnedDb | tr -d '"'`
+        currPinnedDb=$(echo "$pinnedDBArray" | jq ".[$i]")
+        currPinnedDb=$(echo "$currPinnedDb" | tr -d '"')
 
         i=$((i+1))
- 
-        response=`curl -s -X GET \
-        "https://api.supertokens.io/0/plugin-interface/dependency/plugin/latest?password=$SUPERTOKENS_API_KEY&planType=FREE&mode=DEV&version=$piVersion&pluginName=$currPinnedDb" \
-        -H 'api-version: 0'`
-        if [[ `echo $response | jq .plugin` == "null" ]]
+
+        if [[ $currPinnedDb == "sqlite" ]]
         then
-            echo "fetching latest X.Y version for $currPinnedDb given plugin-interface X.Y version: $piVersion gave response: $response"
-            exit 1
+          # shellcheck disable=SC2034
+          continue=1
+        else
+          response=$(curl -s -X GET \
+          "https://api.supertokens.io/0/plugin-interface/dependency/plugin/latest?password=$SUPERTOKENS_API_KEY&planType=FREE&mode=DEV&version=$piVersion&pluginName=$currPinnedDb" \
+          -H 'api-version: 0')
+          if [[ $(echo "$response" | jq .plugin) == "null" ]]
+          then
+              echo "fetching latest X.Y version for $currPinnedDb given plugin-interface X.Y version: $piVersion gave response: $response"
+              exit 1
+          fi
+          pinnedDbVersionX2=$(echo $response | jq .plugin | tr -d '"')
+
+          response=$(curl -s -X GET \
+          "https://api.supertokens.io/0/plugin/latest?password=$SUPERTOKENS_API_KEY&planType=FREE&mode=DEV&version=$pinnedDbVersionX2&name=$currPinnedDb" \
+          -H 'api-version: 0')
+          if [[ $(echo "$response" | jq .tag) == "null" ]]
+          then
+              echo "fetching latest X.Y.Z version for $currPinnedDb, X.Y version: $pinnedDbVersionX2 gave response: $response"
+              exit 1
+          fi
+          pinnedDbVersionTag=$(echo "$response" | jq .tag | tr -d '"')
+          pinnedDbVersion=$(echo "$response" | jq .version | tr -d '"')
+          ./startDb.sh "$currPinnedDb"
         fi
-        pinnedDbVersionX2=$(echo $response | jq .plugin | tr -d '"')
-                
-        response=`curl -s -X GET \
-        "https://api.supertokens.io/0/plugin/latest?password=$SUPERTOKENS_API_KEY&planType=FREE&mode=DEV&version=$pinnedDbVersionX2&name=$currPinnedDb" \
-        -H 'api-version: 0'`
-        if [[ `echo $response | jq .tag` == "null" ]]
-        then
-            echo "fetching latest X.Y.Z version for $currPinnedDb, X.Y version: $pinnedDbVersionX2 gave response: $response"
-            exit 1
-        fi
-        pinnedDbVersionTag=$(echo $response | jq .tag | tr -d '"')
-        pinnedDbVersion=$(echo $response | jq .version | tr -d '"')
-        ./startDb.sh $currPinnedDb
+
         cd ../../
         git clone git@github.com:supertokens/supertokens-root.git
         cd supertokens-root
         coreX=$(cut -d'.' -f1 <<<"$coreVersion")
         coreY=$(cut -d'.' -f2 <<<"$coreVersion")
-        echo -e "core,$coreX.$coreY\nplugin-interface,$piVersion\n$currPinnedDb-plugin,$pinnedDbVersionX2" > modules.txt
+        if [[ $currPinnedDb == "sqlite" ]]
+        then
+          echo -e "core,$coreX.$coreY\nplugin-interface,$piVersion" > modules.txt
+        else
+          echo -e "core,$coreX.$coreY\nplugin-interface,$piVersion\n$currPinnedDb-plugin,$pinnedDbVersionX2" > modules.txt
+        fi
         ./loadModules
         cd supertokens-core
         git checkout dev-v$coreVersion
         cd ../supertokens-plugin-interface
         git checkout $currTag
-        cd ../supertokens-$currPinnedDb-plugin
-        git checkout $pinnedDbVersionTag
+        if [[ $currPinnedDb == "sqlite" ]]
+        then
+          # shellcheck disable=SC2034
+          continue=1
+        else
+          cd ../supertokens-$currPinnedDb-plugin
+          git checkout $pinnedDbVersionTag
+        fi
         cd ../
         echo $SUPERTOKENS_API_KEY > apiPassword
         ./startTestingEnv --cicd
@@ -134,29 +151,43 @@ do
         fi
         cd ../
         rm -rf supertokens-root
-        curl -o supertokens.zip -s -X GET \
-        "https://api.supertokens.io/0/app/download?pluginName=$currPinnedDb&os=linux&mode=DEV&binary=FREE&targetCore=$coreVersion&targetPlugin=$pinnedDbVersion" \
-        -H 'api-version: 0'
-        unzip supertokens.zip -d .
-        rm supertokens.zip
-        cd supertokens
-        ../project/.circleci/testCli.sh
-        if [[ $? -ne 0 ]]
+
+        if [[ $currPinnedDb == "sqlite" ]]
         then
-            echo "cli testing failed... exiting!"
-            exit 1
+          # shellcheck disable=SC2034
+          continue=1
+        else
+          curl -o supertokens.zip -s -X GET \
+          "https://api.supertokens.io/0/app/download?pluginName=$currPinnedDb&os=linux&mode=DEV&binary=FREE&targetCore=$coreVersion&targetPlugin=$pinnedDbVersion" \
+          -H 'api-version: 0'
+          unzip supertokens.zip -d .
+          rm supertokens.zip
+          cd supertokens
+          ../project/.circleci/testCli.sh
+          if [[ $? -ne 0 ]]
+          then
+              echo "cli testing failed... exiting!"
+              exit 1
+          fi
+          cd ../
         fi
-        cd ../
+
         rm -rf supertokens
         cd project/.circleci
-        ./stopDb.sh $currPinnedDb
+        if [[ $currPinnedDb == "sqlite" ]]
+        then
+          # shellcheck disable=SC2034
+          continue=1
+        else
+          ./stopDb.sh $currPinnedDb
+        fi
     done
 done 10<pluginInterfaceExactVersionsOutput
 
 if [[ $someTestsRan = "true" ]]
 then
     echo "calling /core PATCH to make testing passed"
-    responseStatus=`curl -s -o /dev/null -w "%{http_code}" -X PATCH \
+    responseStatus=$(curl -s -o /dev/null -w "%{http_code}" -X PATCH \
         https://api.supertokens.io/0/core \
         -H 'Content-Type: application/json' \
         -H 'api-version: 0' \
@@ -165,8 +196,8 @@ then
             \"planType\":\"FREE\",
             \"version\":\"$coreVersion\",
             \"testPassed\": true
-        }"`
-    if [ $responseStatus -ne "200" ]
+        }")
+    if [ "$responseStatus" -ne "200" ]
     then
         echo "patch api failed"
         exit 1
